Split proxystore.store.base into submodules (#392)
